Type TOTP remaining seconds

+++ nook-app/nook-platform/nook-wasm/src/types/secret_data.rs
@@ -312,7 +312,8 @@ impl NookTotpCode {
 
     #[allow(clippy::cast_precision_loss)]
     pub(crate) fn from_core(value: nook_core::TotpCode, unix_seconds: u64) -> Self {
-        let seconds_remaining = u32::try_from(value.seconds_remaining).unwrap_or(u32::MAX);
+        let seconds_remaining =
+            u32::try_from(u64::from(value.seconds_remaining)).unwrap_or(u32::MAX);
         Self {
             code: value.code,
             seconds_remaining,
